+++ src/commands/Setup.js
@@ -1,16 +1,16 @@
 'use strict'
 import Ansible from '../Ansible'
 
-class Restart extends Ansible {
+class Setup extends Ansible {
   _gatherArgs (cargo, cb) {
     super._gatherArgs(cargo, (err, args) => {
       if (err) {
         return cb(err)
       }
-      args.push(`${this.runtime.config.global.restart_file}`)
+      args.push(`${this.runtime.config.global.setup_file}`)
       return cb(null, args)
     })
   }
 }
 
-module.exports = Restart
+module.exports = Setup